remove `None` variants from `Action`s and use `Option` instead
